Move backend funcions to config.php
